Add save and reset functionality to PauseMenu; update MeshData and WorkerInternalMeshData to use Uint8Array

+++ src/code/World/Chunk/DataStructures/WorkerInternalMeshData.ts
@@ -2,7 +2,7 @@ import { ResizableTypedArray } from "./ResizableTypedArray";
 
 export type WorkerInternalMeshData = {
   faceDataA: ResizableTypedArray<Uint8Array>;
-  faceDataB: ResizableTypedArray<Uint16Array>;
+  faceDataB: ResizableTypedArray<Uint8Array>;
   faceDataC: ResizableTypedArray<Uint8Array>;
   faceCount: number;
 };
